pyproject.toml

[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
pymgclient = "^0.1.0"
flask_dance = "^3.0.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
